Converted BytePixel and UShortPixel to UIntPixel to resolve problems with reading in legacy raster files.
git-svn-id: http://Marc-PC/svn/Full@256 85f15e0e-5137-b144-a863-98bd542365a3

+++ trunk/base-harvest/trunk/src/IManagementAreaDataset.cs
@@ -14,6 +14,6 @@ namespace Landis.Extension.BaseHarvest
         /// <returns>
         /// null if there is no management area with the specified map code.
         /// </returns>
-        ManagementArea Find(ushort mapCode);
+        ManagementArea Find(uint mapCode);
     }
 }
